requirements.txt
torch
torchvision
pillow
streamlit
matplotlib
pytesseract
transformers
opencv-python
ultralytics
numpy
pandas
matplotlib
